% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbTools.R
\name{Stem}
\alias{Stem}
\title{Stem}
\usage{
Stem(x, y, pch = 16, linecol = 1, clinecol = 1, ...)
}
\arguments{
\item{x}{horizontal-axis values}

\item{y}{vertical-axis values}

\item{pch}{integer value, style of points (\code{pch = 21}: circle without fill, see \code{plot} \code{pch} parameter)}

\item{linecol}{color of the plot}

\item{clinecol}{zero axis color}

\item{...}{other parameters passed to \code{plot} function}
}
\value{
creates a plot in base plotting system.
}
\description{
Matlab-like stem plotting function for discrete series.
}
\details{
Discrete plots using base plotting system.

Author: Matti Pastell, Sep 11 2009
http://mpastell.com/2009/09/11/matlab-style-stem-plot-with-r/
}
\examples{
t <- seqM(from = 0, to = 2*pi, length.out = 20)
Stem(t, sin(t))
Stem(t, sin(t), pch=21)
Stem(t, sin(t), pch=21, line="blue")
Stem(t, sin(t), main = "Default style")
}
\seealso{
For interactive time-series plots, see package \code{dygraphs}.
}

