% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbTools-package.R
\docType{package}
\name{tbTools}
\alias{tbTools}
\alias{tbTools-package}
\title{tbTools}
\description{
Tomas' personal mix of utilities
}
\details{
Mix of things that I missed in R. Matlab-like dot operator, stem plot (base plotting system), round2 with order, ifft etc.

\code{\link{seqM}} Matlab-like behaviour of colon operator or linspace for creating sequences, for-loop friendly.
\code{\link{round2}} Rounds a number to the specified order. Round half away from zero (this is the difference from built-in \code{round} function.)
\code{\link{ifft}} Inverse Fast Fourier Transform (discrete FT), Matlab-like behavior.

\code{\link{Stem}} Matlab-like stem plotting function for discrete series.

\code{\link{isInt}} Returns TRUE / FALSE whether it is exactly 1 integer number (in fact, the class can be numeric but the number must be integer), non-missing
\code{\link{isNum}} Returns TRUE / FALSE whether it is exactly 1 number (numeric or integer vector of length 1, non-missing)
\code{\link{isString}} Returns TRUE / FALSE whether it is exactly 1 character string (character vector of length 1, non-missing)

\code{\link{strTrim}} Trim leading and trailing whitespace in character string. Way faster than str_trim() or trimws().

\code{\link{str_contains}} Find string in another string (without regular expressions), returns TRUE / FALSE.
\code{\link{str_find}} Find string in another string (without regular expressions), returns indices of all occurences.
\code{\link{str_find1}} Find string in another string (without regular expressions), returns indices of the first occurence only.
}

